+++ docs/source/changes.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.0.6 (not released)
+--------------------
+
+Changes :
+
+- Added single-var groupings.
+- Added props method to `BenchmarkObj` to list meta information about it.
+
+
 0.0.5 (2020-10-15)
 ------------------